refactor: refac. components location (header)
