+++ src/components/header/header-icons/header-icons.component.spec.ts
