+++ src/components/menu/menu-item/menu-item.component.scss
@@ -1,5 +1,4 @@
 .app-menu__button {
-  // text-decoration: none;
   padding: 0.5rem 1rem;
   border-radius: var(--button-radius-padding);
 }
