+++ src/components/menu/menu.component.scss
@@ -20,8 +20,4 @@
       state-layer-color: var(--mat-sys-primary),
       ripple-color: var(--mat-sys-inverse-primary)));
   color: var(--color-primary);
-  // padding: 0 var(--button-radius-padding, 0.5rem);
-  // &>span.mat-mdc-button-persistent-ripple.mdc-icon-button__ripple {
-  //   border-radius: var(--button-radius-padding, 0.5rem);
-  // }
 }
