+++ src/constants/menu-items.ts
@@ -1,7 +1,6 @@
 import { MenuItem } from '../models/types';
 
 export const MENU_ITEMS: MenuItem[] = [
-  { item: 'Home', route: 'main' },
   { item: 'Catalog', route: 'catalog' },
   { item: 'Search', route: 'search' },
   { item: 'Contact Us', route: 'contact' },
